requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
SQLAlchemy>=2.0
psycopg2-binary
alembic
passlib[bcrypt]
pyjwt
python-multipart
pandas
python-dateutil
openpyxl
redis
APScheduler
google-auth
google-auth-oauthlib
gspread
# Optional drivers (install as needed):
# pymysql
# pyodbc
# cx_Oracle
# snowflake-connector-python
# snowflake-sqlalchemy
# bigquery-sqlalchemy
